+++ src/main/java/com/awesomehippo/betterzoom/config/Config.java
@@ -6,14 +6,14 @@ import net.neoforged.neoforge.common.ModConfigSpec;
 public class Config {
     private static final ModConfigSpec.Builder BUILDER = new ModConfigSpec.Builder();
 
-    public static final double MIN_ZOOM_INCREMENT = 0.1;
-    public static final double MAX_ZOOM_INCREMENT = 20.0;
+    public static final int MIN_ZOOM_INCREMENT = 1;
+    public static final int MAX_ZOOM_INCREMENT = 20;
     public static final double MIN_SENSITIVITY = 0.01;
     public static final double MAX_SENSITIVITY = 1.0;
     public static final double MIN_SMOOTH_EASE = 0.05;
     public static final double MAX_SMOOTH_EASE = 0.2;
 
-    public static final ModConfigSpec.DoubleValue ZOOM_STEP;
+    public static final ModConfigSpec.IntValue ZOOM_STEP;
     public static final ModConfigSpec.DoubleValue ZOOM_SENSITIVITY_MULTIPLIER;
     public static final ModConfigSpec.BooleanValue HOLD_TO_ZOOM;
     public static final ModConfigSpec.DoubleValue SMOOTH_EASING_FACTOR;
@@ -42,7 +42,7 @@ public class Config {
     static {
         BUILDER.comment("Better Zoom Configuration");
 
-        ZOOM_STEP = BUILDER.translation("betterzoom.config.zoomstep.label").comment("The zoom increment step per zoom in/out (higher values zoom/dezoom faster)").defineInRange("zoomStep",5.0, MIN_ZOOM_INCREMENT, MAX_ZOOM_INCREMENT);
+        ZOOM_STEP = BUILDER.translation("betterzoom.config.zoomstep.label").comment("The zoom increment step per zoom in/out (higher values zoom/dezoom faster)").defineInRange("zoomStep", 5, MIN_ZOOM_INCREMENT, MAX_ZOOM_INCREMENT);
         ZOOM_SENSITIVITY_MULTIPLIER = BUILDER.translation("betterzoom.config.sensitivity.label").comment("Mouse sensitivity multiplier while zooming (lower values make it less sensitive)").defineInRange("zoomSensitivity", 0.01, MIN_SENSITIVITY, MAX_SENSITIVITY);
         AUTO_ADJUST_SENSITIVITY = BUILDER.translation("betterzoom.config.autosensitivity.label").comment("Adjusts mouse sensitivity based on the zoom level").define("autoAdjustSensitivity", true);
         HOLD_TO_ZOOM = BUILDER.translation("betterzoom.config.hold.label").comment("Require holding the zoom key").define("holdToZoom", true);